+++ src/app/api/listObjects/route.ts
@@ -1,0 +1,22 @@
+import { env } from "@/app/env.mjs";
+import { S3Client } from "@aws-sdk/client-s3";
+import { NextResponse } from "next/server";
+import { listContents } from "@/utils/apiHelper";
+
+export async function GET(request: Request) {
+  const url = new URL(request.url);
+
+  const prefix = url.searchParams.get("prefix");
+  console.log("prefix", prefix);
+
+  const s3Client = new S3Client({
+    region: env.AWS_REGION,
+    credentials: {
+      accessKeyId: env.AWS_ACCESS_KEY_ID,
+      secretAccessKey: env.AWS_SECRET_ACCESS_KEY,
+    },
+  });
+
+  const data = await listContents({ prefix: prefix ? prefix : "", s3Client });
+  return NextResponse.json({ data });
+}
