+++ src/components/tablecomponents/user-nav.tsx
@@ -13,11 +13,12 @@ import { PlusCircle } from "lucide-react";
 import { useState } from "react";
 
 interface UserNavProps {
+  orgSlug: string;
   username: string;
   userId: string;
 }
 
-export function UserNav({ username, userId }: UserNavProps) {
+export function UserNav({ username, userId, orgSlug }: UserNavProps) {
   const [isOpen, setIsOpen] = useState(false);
   return (
     <Dialog open={isOpen} onOpenChange={setIsOpen}>
@@ -35,7 +36,12 @@ export function UserNav({ username, userId }: UserNavProps) {
             and use.
           </DialogDescription>
         </DialogHeader>
-        <DropZone username={username} userId={userId} setIsOpen={setIsOpen} />
+        <DropZone
+          orgSlug={orgSlug}
+          username={username}
+          userId={userId}
+          setIsOpen={setIsOpen}
+        />
       </DialogContent>
     </Dialog>
   );
